requirements.txt
chromadb
pandas
tqdm
openai
requests
fastapi
uvicorn
python-multipart
pydantic
python-dotenv







